Issue #40: Extract an Eclipse-independent jar
Eclipse-related content got removed from org.eclipse.tm4e.core, and
placed in a new org.eclipse.tm4e.registry instead.
So org.eclipse.tm4e.core should be usable and fully working as a Maven
artifact.
