+++ org.eclipse.tm4e.core/META-INF/MANIFEST.MF
@@ -3,12 +3,11 @@ Bundle-ManifestVersion: 2
 Bundle-Name: %pluginName
 Bundle-Vendor: %providerName
 Bundle-Localization: plugin
-Bundle-SymbolicName: org.eclipse.tm4e.core;singleton:=true
+Bundle-SymbolicName: org.eclipse.tm4e.core
 Bundle-Version: 0.1.0.qualifier
 Require-Bundle: org.apache.batik.css;bundle-version="1.7.0";resolution:=optional,
  org.apache.batik.util;bundle-version="1.7.0";resolution:=optional,
  com.google.gson;resolution:=optional,
- org.eclipse.core.runtime,
  org.jcodings;bundle-version="1.0.18",
  org.joni;bundle-version="2.1.11"
 Import-Package: org.w3c.css.sac;resolution:=optional,
@@ -16,10 +15,10 @@ Import-Package: org.w3c.css.sac;resolution:=optional,
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Export-Package: org.eclipse.tm4e.core,
  org.eclipse.tm4e.core.grammar,
- org.eclipse.tm4e.core.grammars,
+ org.eclipse.tm4e.core.internal.grammars;x-internal:=true,
  org.eclipse.tm4e.core.internal.matcher;x-friends:="org.eclipse.tm4e.core.tests",
+ org.eclipse.tm4e.core.internal.types;x-internal:=true,
  org.eclipse.tm4e.core.model,
  org.eclipse.tm4e.core.registry,
  org.eclipse.tm4e.core.theme,
  org.eclipse.tm4e.core.theme.css
-Bundle-ActivationPolicy: lazy
