+++ org.eclipse.tm4e.core/build.properties
@@ -2,6 +2,4 @@ source.. = src/main/java/,\
            src/main/resources/
 bin.includes = META-INF/,\
                .,\
-               plugin.properties,\
-               schema/,\
-               plugin.xml
+               plugin.properties
