+++ org.eclipse.tm4e.core/src/main/java/org/eclipse/tm4e/core/internal/grammars/SyncRegistry.java
@@ -13,7 +13,7 @@
  *  - Microsoft Corporation: Initial code, written in TypeScript, licensed under MIT license
  *  - Angelo Zerr <angelo.zerr@gmail.com> - translation and adaptation to Java
  */
-package org.eclipse.tm4e.core.registry;
+package org.eclipse.tm4e.core.internal.grammars;
 
 import java.util.ArrayList;
 import java.util.Collection;
