--- org.eclipse.tm4e.core/src/main/java/org/eclipse/tm4e/core/registry/IGrammarInfo.java
@@ -1,27 +1,0 @@
-/**
- *  Copyright (c) 2015-2017 Angelo ZERR.
- *  All rights reserved. This program and the accompanying materials
- *  are made available under the terms of the Eclipse Public License v1.0
- *  which accompanies this distribution, and is available at
- *  http://www.eclipse.org/legal/epl-v10.html
- *
- * Initial code from https://github.com/Microsoft/vscode-textmate/
- * Initial copyright Copyright (C) Microsoft Corporation. All rights reserved.
- * Initial license: MIT
- *
- * Contributors:
- *  - Microsoft Corporation: Initial code, written in TypeScript, licensed under MIT license
- *  - Angelo Zerr <angelo.zerr@gmail.com> - translation and adaptation to Java
- */
-package org.eclipse.tm4e.core.registry;
-
-public interface IGrammarInfo {
-
-	String[] getFileTypes();
-
-	String getName();
-
-	String getScopeName();
-
-	String firstLineMatch();
-}
