+++ org.eclipse.tm4e.core/src/main/java/org/eclipse/tm4e/core/registry/Registry.java
@@ -25,6 +25,7 @@ import java.util.List;
 import org.eclipse.tm4e.core.TMException;
 import org.eclipse.tm4e.core.grammar.IGrammar;
 import org.eclipse.tm4e.core.internal.grammar.reader.GrammarReader;
+import org.eclipse.tm4e.core.internal.grammars.SyncRegistry;
 import org.eclipse.tm4e.core.internal.types.IRawGrammar;
 
 public class Registry {
