+++ org.eclipse.tm4e.core/src/test/java/org/eclipse/tm4e/core/internal/grammar/parser/json/JSONPListParserTest.java
@@ -10,7 +10,6 @@
  */
 package org.eclipse.tm4e.core.internal.grammar.parser.json;
 
-import org.eclipse.tm4e.core.internal.grammar.parser.json.JSONPListParser;
 import org.eclipse.tm4e.core.internal.grammar.reader.IGrammarParser;
 import org.eclipse.tm4e.core.internal.types.IRawGrammar;
 import org.junit.Test;
