+++ org.eclipse.tm4e.markdown/META-INF/MANIFEST.MF
@@ -7,6 +7,7 @@ Bundle-SymbolicName: org.eclipse.tm4e.markdown;singleton:=true
 Bundle-Version: 0.1.0.qualifier
 Require-Bundle: org.eclipse.core.runtime,
  org.eclipse.tm4e.core,
+ org.eclipse.tm4e.registry,
  org.eclipse.tm4e.ui
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Export-Package: org.eclipse.tm4e.markdown,
