+++ org.eclipse.tm4e.registry/.gitignore
@@ -1,0 +1,1 @@
+/bin/
