+++ org.eclipse.tm4e.registry/META-INF/MANIFEST.MF
@@ -1,0 +1,12 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Registry management for TextMate grammars
+Bundle-SymbolicName: org.eclipse.tm4e.registry;singleton:=true
+Bundle-Version: 0.1.0.qualifier
+Bundle-Vendor: %providerName
+Bundle-RequiredExecutionEnvironment: JavaSE-1.8
+Require-Bundle: org.eclipse.tm4e.core;bundle-version="0.1.0",
+ org.eclipse.core.runtime;bundle-version="3.11.0"
+Export-Package: org.eclipse.tm4e.registry.internal
+Bundle-Activator: org.eclipse.tm4e.registry.internal.TMEclipseRegistryPlugin
+Bundle-Localization: plugin
