+++ org.eclipse.tm4e.registry/build.properties
@@ -1,0 +1,7 @@
+source.. = src/main/java
+output.. = bin/
+bin.includes = META-INF/,\
+               schema/,\
+               plugin.xml,\
+               plugin.properties,\
+               .
