+++ org.eclipse.tm4e.registry/plugin.properties
@@ -8,5 +8,8 @@
 # Contributors:
 #     Angelo Zerr <angelo.zerr@gmail.com> - Initial API and implementation 
 ###############################################################################
-pluginName=TextMate support in Java
+pluginName=TextMate support in Eclipse IDE - Registry management
 providerName=Eclipse TM4E
+
+# Extension Points
+Grammars.extension.name=TextMate grammars provider extension point.
