+++ org.eclipse.tm4e.registry/plugin.xml
