+++ org.eclipse.tm4e.registry/pom.xml
@@ -1,0 +1,10 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <artifactId>org.eclipse.tm4e.registry</artifactId>
+  <packaging>eclipse-plugin</packaging>
+  <parent>
+    <groupId>org.eclipse</groupId>
+    <artifactId>org.eclipse.tm4e</artifactId>
+    <version>0.1.0-SNAPSHOT</version>
+  </parent>
+</project>
