+++ org.eclipse.tm4e.registry/src/main/java/org/eclipse/tm4e/registry/internal/IGrammarRegistryManager.java
@@ -8,7 +8,7 @@
  *  Contributors:
  *  Angelo Zerr <angelo.zerr@gmail.com> - initial API and implementation
  */
-package org.eclipse.tm4e.core.grammars;
+package org.eclipse.tm4e.registry.internal;
 
 import org.eclipse.core.runtime.content.IContentType;
 import org.eclipse.tm4e.core.grammar.IGrammar;
