+++ org.eclipse.tm4e.ui.tests/META-INF/MANIFEST.MF
@@ -16,7 +16,8 @@ Require-Bundle: org.eclipse.tm4e.core,
  org.junit,
  org.eclipse.ui.genericeditor;resolution:=optional,
  org.eclipse.ui.ide,
- org.eclipse.ui.tests.harness;bundle-version="1.4.0"
+ org.eclipse.ui.tests.harness;bundle-version="1.4.0",
+ org.eclipse.tm4e.registry;bundle-version="0.1.0"
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-ActivationPolicy: lazy
 Import-Package: org.eclipse.core.filebuffers
