+++ org.eclipse.tm4e.ui.tests/src/test/java/org/eclipse/tm4e/ui/RegistryTest.java
@@ -2,8 +2,8 @@ package org.eclipse.tm4e.ui;
 
 import org.eclipse.core.runtime.Platform;
 import org.eclipse.core.runtime.content.IContentType;
-import org.eclipse.tm4e.core.TMCorePlugin;
 import org.eclipse.tm4e.core.grammar.IGrammar;
+import org.eclipse.tm4e.registry.internal.TMEclipseRegistryPlugin;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -12,7 +12,7 @@ public class RegistryTest {
 	@Test
 	public void testGrammarRegistered() {
 		IContentType contentType = Platform.getContentTypeManager().getContentType("org.eclipse.tm4e.ui.tests.testContentType");
-		IGrammar grammar = TMCorePlugin.getGrammarRegistryManager().getGrammarFor(new IContentType[] { contentType });
+		IGrammar grammar = TMEclipseRegistryPlugin.getGrammarRegistryManager().getGrammarFor(new IContentType[] { contentType });
 		Assert.assertNotNull(grammar);
 	}
 }
