+++ org.eclipse.tm4e.ui/META-INF/MANIFEST.MF
@@ -10,7 +10,8 @@ Require-Bundle: org.eclipse.tm4e.core,
  org.eclipse.core.runtime,
  org.eclipse.ui,
  org.eclipse.core.resources,
- org.eclipse.core.filesystem
+ org.eclipse.core.filesystem,
+ org.eclipse.tm4e.registry;bundle-version="0.0.1"
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Export-Package: org.eclipse.tm4e.ui,
  org.eclipse.tm4e.ui.model,
