+++ pom.xml
@@ -33,6 +33,7 @@
 		<module>target-platform</module>
 		<module>org.eclipse.tm4e.core</module>
 		<module>org.eclipse.tm4e.core.tests</module>
+		<module>org.eclipse.tm4e.registry</module>
 		<module>org.eclipse.tm4e.ui</module>
 		<module>org.eclipse.tm4e.ui.tests</module>
 		<module>org.eclipse.tm4e.markdown</module>
