+++ tm-feature/feature.xml
@@ -27,6 +27,14 @@
          unpack="false"/>
 
    <plugin
+         id="org.eclipse.tm4e.registry"
+         download-size="0"
+         install-size="0"
+         version="0.0.0"
+         unpack="false"/>
+
+
+   <plugin
          id="org.eclipse.tm4e.ui"
          download-size="0"
          install-size="0"
